requirements.txt
aiohttp==3.9.5
fastapi==0.112.2
google-api-python-client==2.130.0
googlemaps
httpx==0.27.0
Jinja2==3.1.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.35
langchain-google-community==1.0.8
langchainhub==0.1.21
langchain-openai==0.1.7
langgraph==0.0.55
langsmith==0.1.105
numpy==1.26.4
openai==1.30.3
opencv-python==4.9.0.80
pandas==2.2.2
pydantic==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.2
uvicorn==0.30.6
pillow==10.4.0
PyYAML==6.0.1
tokenizers==0.19.1
tqdm==4.66.4
transformers==4.41.1